Fixed placement of setWithAuthorizationClass
